+++ CHANGELOG.md
@@ -9,7 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- add classes for Additive, Preventive and Subtractive rules for all entity types
+- add types, lists and lookups for all three rule types to `mex.common.models`
+
 ### Changes
+
+- move aux-extractor documentation from readme to `__init__` to have it in sphinx
+- move `BaseModel` specific descriptions from class to model to avoid duplication
+- BREAKING: move `FILTER_MODEL_BY_EXTRACTED_CLASS_NAME` to `mex.common.models`
+- BREAKING: move `MAPPING_MODEL_BY_EXTRACTED_CLASS_NAME` to `mex.common.models`
+- BREAKING: change `MEX_PRIMARY_SOURCE_IDENTIFIER` to end in `1`,
+  so that it differs from `MEX_PRIMARY_SOURCE_STABLE_TARGET_ID`
 
 ### Deprecated